(modularize components): Implement component separation
- separate Header component
- separate Player component
- separate Counter component
- add named import Component

[Starts #003]
